Modules
Broke down the one file into modules for readability, and upkeep.
