+++ index.html
@@ -3,11 +3,11 @@
 
 <HEAD>
 	<TITLE> Chess Puzzle Maker </TITLE>
-	<script src="chess.js"> </script>
+	<script type="module" src="main.js"> </script>
 	<link rel="stylesheet" type="text/css" href="chess.css">
 </HEAD>
 
-<BODY id='body' onload="main()">
+<BODY id='body'>
 	<H1> Chess Puzzle Maker </H1>
 	<div>
 		<canvas id="chessboard">Canvas not supported.</canvas>
