+++ modules/chess.mjs
@@ -1,232 +1,6 @@
-// chess.js
-
-class BaseClass {
-	#left = 0
-	#right = 0
-	#top = 0
-	#bottom = 0
-	#cx = 0
-	#cy = 0
-	#width = 0 
-	#height = 0
-
-	set width(w) {
-		this.#width = w;
-	}
-
-	get width() {
-		return this.#width;
-	}
-
-	set height(h) {
-		this.#height = h;
-	}
-
-	get height() {
-		return this.#height;
-	}
-
-	set left(l) {
-		this.#left = l;
-	}
-
-	get left() {
-		return this.#left;
-	}
-
-	set right(r) {
-		this.#right = r;
-	}
-
-	get right() {
-		return this.#right;
-	}
-
-	set top(t) {
-		this.#top = t;
-	}
-
-	get top() {
-		return this.#top;
-	}
-
-	set bottom(b) {
-		this.#bottom = b;
-	}
-
-	get bottom() {
-		return this.#bottom;
-	}
-
-	set cx(x) {
-		this.#cx = x;
-	}
-
-	get cx() {
-		return this.#cx;
-	}
-
-	set cy(y) {
-		this.#cy = y;
-	}
-
-	get cy() {
-		return this.#cy;
-	}
-
-	constructor(left, top, width, height) {
-		this.setSize(width, height);
-		this.setPos(left, top);
-	}
-
-	setSize(width, height) {
-		this.width = width;
-		this.height = height;
-	}
-
-	setPos(left, top) {
-		this.left = left;
-		this.right = left + this.width;
-		this.top = top;
-		this.bottom = top + this.height;
-		this.cx = left + this.width / 2;
-		this.cy = top + this.height / 2;
-	}
-}
-
-function DrawImage(context, imageSrc, left, top, width, height) {
-	var image = new Image();
-	image.onload = function() {
-		context.drawImage(image, left, top, width, height);
-	};
-	image.src = imageSrc;
-}
-
-class Sprite extends BaseClass {
-	#isLoaded = false
-	#src = ''
-	#image = null
-
-	constructor(src, left, top, width, height) {
-		super(left, top, width, height);
-		this.#src = src;
-		this.load_image();
-	}
-
-	load_image() {
-		this.#image = new Image();
-		this.#image.onload = () => { this.#isLoaded = true; };
-		this.#image.src = this.#src;
-	}
-
-	draw(ctx) {
-		if (this.#isLoaded) {
-			ctx.drawImage(this.#image, this.left, this.top, this.width, this.height);
-		}
-	}
-}
-
-class Text {
-	#string = ''
-	#cx = 0
-	#cy = 0
-	#font = ''
-
-	get string() {
-		return this.#string;
-	}
-
-	set string(value) {
-		this.#string = value;
-	}
-
-	constructor(string, cx, cy, font='20px Arial') {
-		this.#cx = cx;
-		this.#cy = cy;
-		this.#string = string;
-		this.#font = font;
-	}
-
-	draw(ctx) {
-		ctx.font = this.#font;
-		ctx.textAlign = 'center';
-		ctx.textBaseline = 'middle';
-		ctx.fillText(this.#string, this.#cx, this.#cy);
-	}
-}
-
-class Rect extends BaseClass {
-	#color = '#FF0000'
-
-	constructor(left, top, width, height, color) {
-		super(left, top, width, height);
-		this.#color = color;
-	}
-
-	draw(ctx) {
-		ctx.beginPath();
-		ctx.strokeStyle = this.#color;
-		ctx.rect(this.left, this.top, this.width, this.height);
-		ctx.stroke();
-	}
-}
-
-const EventTypes = {
-	MOUSE_DOWN : 'mouse_down',
-	MOUSE_UP   : 'mouse_up'
-}
-
-class EventSystem {
-	#canvasLeft = 0
-	#canvasTop = 0 
-	#delegates = {}
-	#keys = []
-
-	get delegates() {
-		return this.#delegates;
-	}
-
-	constructor() {
-		for (var type in EventTypes) {
-			this.delegates[EventTypes[type]] = []
-		}
-	}
-
-	wire_events(canvas) {
-		var rect = canvas.getBoundingClientRect();
-		this.#canvasLeft = rect.left;
-		this.#canvasTop = rect.top;
-		canvas.onmousedown = (mouseEventArgs) => this.onMouseDown(mouseEventArgs);
-		canvas.onmouseup = (mouseEventArgs) => this.onMouseUp(mouseEventArgs);
-	}
-
-	add_listener(eventType, func) {
-		this.delegates[eventType].push(func);
-	}
-
-	callAllDelegateOfType(eventType, eventArgs) {
-		for (var func of this.delegates[eventType]) {
-			func(eventArgs);
-		}
-	}
-
-	onMouseDown(mouseEventArgs) {
-		this.convertScreenToCanvasPoint(mouseEventArgs);
-		this.callAllDelegateOfType(EventTypes.MOUSE_DOWN, mouseEventArgs);
-	}
-
-	onMouseUp(mouseEventArgs) {
-		this.convertScreenToCanvasPoint(mouseEventArgs);
-		this.callAllDelegateOfType(EventTypes.MOUSE_UP, mouseEventArgs);
-	}
-
-	convertScreenToCanvasPoint(eventArgs) {
-		eventArgs.canvasX = eventArgs.clientX - this.#canvasLeft;
-		eventArgs.canvasY = eventArgs.clientY - this.#canvasTop;
-	}
-}
-
-let events = new EventSystem();
+//chess.mjs
+import { Events, EventTypes } from './events.mjs';
+import { Rect, Text, Sprite } from './drawing.mjs';
 
 const ChessInfo = {
 	SPRITE_FOLDER     : 'sprites',
@@ -405,7 +179,7 @@ class Chessboard {
 	}
 
 	wire_events() {
-		events.add_listener(EventTypes.MOUSE_DOWN, (mouseEventArgs) => this.onMouseDown(mouseEventArgs));
+		Events.add_listener(EventTypes.MOUSE_DOWN, (mouseEventArgs) => this.onMouseDown(mouseEventArgs));
 		document.getElementById('btnReset').onclick = (btnEventArgs) => this.onResetClick(btnEventArgs);
 	}
 
@@ -507,44 +281,4 @@ class Chessboard {
 	}
 }
 
-class Runtime { 
-	#isRunning = false
-	#ctx = null
-	#tileSize = 0
-	#chessboard = null
-
-	constructor(ctx, tileSize) {
-		this.#ctx = ctx;
-		this.#tileSize = tileSize;
-		this.#chessboard = new Chessboard(tileSize);
-		this.#isRunning = true;
-	}
-
-	update() {
-
-	}
-
-	draw() {
-		this.#chessboard.draw(this.#ctx);
-	}
-
-	run() {
-		if (this.#isRunning) { 
-			this.update();
-			this.draw();
-			window.requestAnimationFrame(() => { this.run(); });
-		}
-	}
-}
-
-function main() {
-	var canvas = document.getElementById('chessboard');
-	var ctx = canvas.getContext('2d');
-	events.wire_events(canvas);
-	canvasSize = ChessInfo.CHESSBOARD_ROWS * ChessInfo.DEFAULT_TILE_SIZE + 2; //Add 2 for selection border
-	ctx.canvas.width = canvasSize;
-	ctx.canvas.height = canvasSize;
-	var runtime = new Runtime(ctx, ChessInfo.DEFAULT_TILE_SIZE);
-	runtime.run();
-};
-
+export { ChessInfo, Chessboard };
